+++ index.js
@@ -2,7 +2,8 @@ module.exports = {
   compile: require('./lib'),
   generators: {
     csharp: require('./lib/stringify/csharp'),
-    kotlin: require('./lib/stringify/kotlin')
+    kotlin: require('./lib/stringify/kotlin'),
+    typescript: require('./lib/stringify/typescript')
   },
   parsers: {
     javascript: require('./lib/parse/javascript'),